cleanup and updates
