+++ pyproject.toml
@@ -6,6 +6,7 @@ build-backend = "setuptools.build_meta"
 name = "lofar_facet_selfcal"
 version = "11.3.0"
 description = "LOFAR facet selfcal"
+requires-python = ">=3.9"
 
 [project.scripts]
 h5_merger = "submods.h5_merger:main"
